+++ pyproject.toml
@@ -6,8 +6,10 @@ readme = "README.md"
 requires-python = "==3.12.3"
 dependencies = [
     "flask>=3.1.0",
+    "geoalchemy2>=0.17.1",
     "psycopg2-binary>=2.9.10",
     "python-dotenv>=1.0.1",
     "requests>=2.32.3",
+    "sqlalchemy>=2.0.40",
     "waitress>=3.0.2",
 ]
