+++ server/static/css/admin.css
