+++ server/static/js/admin.js
