+++ server/static/js/common.js
