+++ server/static/js/location.js
