+++ server/static/js/uploadSticker.js
