--- templates/home.html
@@ -1,105 +1,0 @@
-<html>
-    <head>
-        <title>Plakplaats | Sticky Sticker Map</title>
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="stylesheet" href="/static/css/home.css">
-        <link rel="stylesheet" href="https://unpkg.com/leaflet@1.7.1/dist/leaflet.css"
-	    integrity="sha512-xodZBNTC5n17Xt2atTPuE1HxjVMSvLVW9ocqUKLsCC5CXdbqCmblAshOMAS6/keqq/sMZMZ19scR4PsZChSR7A=="
-	    crossorigin=""/>
-        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/leaflet.locatecontrol/dist/L.Control.Locate.min.css" />
-        <link rel="icon" href="static/img/favicon.ico" type="image/ico"/>
-        <link rel="shortcut icon" href="static/img/favicon.ico" type="image/x-icon"/>
-    </head>
-    <body>
-        <div class="navBar">
-            <img src="/static/img/logo_compact_outline_wit.png" height="60px" style="padding-left: 20px;">
-            <!-- <button class="nearYouButton">Stickers near You</button> -->
-            <svg class="addIcon" viewBox="0 0 100 100" width="30px" height="30px" xmls="http://www.w3.org/2000/svg">
-                <line x1="10" y1="50" x2="90" y2="50" stroke="white" stroke-width="15px" stroke-linecap="round"/> 
-                <line x1="50" y1="10" x2="50" y2="90" stroke="white" stroke-width="15px" stroke-linecap="round"/>
-            </svg>
-        </div>
-
-	        <div id="map" ></div>  <!-- Will be filled by leaflet code, resulting in the map -->
-
-        <div id="mapOverlay">
-            <div id="overlayText">Click anywhere to pick location</div>
-        </div>
-        <div class="addContainer">
-            <div class="addView">
-                <span class="addViewTitle">Add a sticker</span>
-                <div class="locationContainer" onclick="setPicker(true)">
-                    <svg class="locationContainerSpinner" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" style="shape-rendering: auto;" width="40px" height="40px" viewBox="0 0 100 100" preserveAspectRatio="xMidYMid">
-                        <circle class="locationContainerCircle" cx="50" cy="50" fill="none" stroke="#e15b64" stroke-width="10" r="35" stroke-dasharray="164.93361431346415 56.97787143782138">
-                            <animateTransform attributeName="transform" type="rotate" repeatCount="indefinite" dur="1s" values="0 50 50;360 50 50" keyTimes="0;1"></animateTransform>
-                        </circle>
-                    </svg>
-                    <svg class="locationIcon" viewBox="0 0 100 100" width="40px" height="40px" xmlns="http://www.w3.org/2000/svg">
-                        <circle class="outerCircle circle" cx="50" cy="50" r="20" fill="blue"></circle>
-                        <circle class="circle" cx="50" cy="50" r="20" fill="blue" stroke="white" stroke-width="3px"></circle>   
-                      </svg>
-                    <span class="locationContainerText"></span>
-                </div>
-                <span class="addInputDes locationInputDes">Latitude</span>
-                <input type="number" class="addInput" id="latitudeInput">
-                <span class="addInputDes locationInputDes">Longitude</span>
-                <input type="number" class="addInput" id="longitudeInput">
-                <div class="addImageInput" onclick="onClickImage()">
-                    <span class="addImageText">Click to upload</span>
-                    <img class="addImagePreview" src="https://images.dog.ceo/breeds/terrier-norwich/n02094258_125.jpg">
-                </div>
-                <div class="boarYearInput">
-                    <label for="boardYearInputSelect">Board year:</label>
-                    <select id="boardYearInputSelect" name="boardYearInputSelect">
-                        <option value="1">B1</option>
-                        <option value="2">B2</option>
-                        <option value="3">B3</option>
-                        <option value="4">B4</option>
-                        <option value="5">B5</option>
-                        <option value="6">B6</option>
-                        <option value="7">B7</option>
-                        <option value="8">B8</option>
-                        <option value="9">B9</option>
-                        <option value="10">B10</option>
-                        <option value="11">B11</option>
-                        <option value="12">B12</option>
-                        <option value="13">B13</option>
-                        <option value="14">B14</option>
-                        <option value="15">B15</option>
-                        <option value="16">B16</option>
-                        <option value="17">B17</option>
-                        <option value="18">B18</option>
-                        <option value="19" selected>B19</option>
-                    </select>
-                </div>
-                <div class="addLogoSelector">
-                </div>
-
-                <span class="addSubmitButton" onclick="submit()">Submit</span>
-            </div>
-            <div class="successView">
-                <span class="addViewTitle">Success</span>
-                <span class="successText">Thanks for submitting your sticker! Your sticker will be visible after it is reviewed. You can keep adding stickers while we are doing this.</span>
-                <span class="successText">Want to know when your sticker is approved? Fill in your email below and receive status updates!</span>
-                <input type="email" class="successEmail" placeholder="Email">
-                <span class="addSubmitButton" onclick="updateEmail()">Submit</span>
-            </div>
-        </div>
-    </body>
-
-     <!-- Make sure you put this AFTER Leaflet's CSS -->
-    <script src="https://unpkg.com/leaflet@1.7.1/dist/leaflet.js"
-        integrity="sha512-XQoYMqMTK8LvdxXYG3nZ448hOEQiglfqkJs1NOQV44cWnUrBc8PkAOcXy20w0vlaXaVUearIOBhiXZ5V3ynxwA=="
-	crossorigin=""></script>
-    <script src="https://cdn.jsdelivr.net/npm/dayjs@1/dayjs.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/dayjs@1/plugin/relativeTime.js"></script>
-    <script>dayjs.extend(window.dayjs_plugin_relativeTime)</script>
-    <script src="https://cdn.jsdelivr.net/npm/leaflet.locatecontrol/dist/L.Control.Locate.min.js" charset="utf-8"></script>
-    <script src="/static/js/common.js"></script>
-    <script src="/static/js/map.js"></script>
-    <script src="/static/js/UI.js"></script>
-    <script src="/static/js/location.js"></script>
-    <script src="/static/js/uploadSticker.js"></script>
-    <script src="/static/js/nearYouOverlay.js"></script>
-    <script src="/static/js/main.js"></script>
-</html>
